Use size_t instead of int in some cases to avoid strlen narrowing conversions

+++ syllabifier_result.hpp
@@ -7,7 +7,7 @@ struct SyllabificationResult {
     char word[kResultLength];
     int numSyllables;
     SyllabificationResult *next;
-    int lastPlace;
+    size_t lastPlace;
 
     SyllabificationResult();
